requirements.txt
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.34
langchain_text_splitters==0.2.2
unstructured[md]==0.11.8
faiss-cpu==1.8.0
transformers==4.36.2
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.2.2